+++ src/membrain_seg/training/__init__.py
@@ -1,3 +1,1 @@
-"""Neural networks implemented as pytorch lightning modules."""
-__all__ = ["SemanticSegmentationUnet"]
-from membrain_seg.training.unet import SemanticSegmentationUnet
+"""Empty init."""
